+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Version 2.2.2
+- Improved Autorec Update Menu and fixed a major bug
+  - It now sorts entries by label
+  - It now shows the type of entry (melee, ranged, ontoken, etc.)
+  - Fixed a bug where the auto-deletion process would error on (and as such, practically delete) animations without metaData
+
 # Version 2.2.1
 - Added removal of outdated effects in the Autorecognition Update Menu.
 - Added Black Tentacles attack animation.